调整标签管理方案

+++ collection-console/src/main/java/com/keem/kochiu/collection/data/bo/ResInfoBo.java
@@ -8,7 +8,7 @@ import javax.validation.constraints.NotNull;
 public class ResInfoBo {
 
     @NotNull(message = "resourceId不能为空")
-    private Integer resourceId;
+    private Long resourceId;
     private String title;
     private String description;
     private Integer star;
